// src/app/api/model/accountDetails.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountDetailsAvatar } from './accountDetailsAvatar';


export interface AccountDetails { 
    avatar?: AccountDetailsAvatar;
    id?: number;
    iso_639_1?: string;
    iso_3166_1?: string;
    name?: string;
    include_adult?: boolean;
    username?: string;
}